合計欄をformから削除、表示処理の変更
