+++ src/pages/projEstimate/SummaryTable/DisplayCellContent.tsx
@@ -1,16 +1,17 @@
 import { Typography } from '@mui/material';
-import { useField } from 'formik';
 
-export type DisplayCellContentProps = {
+
+const DisplayCellContent = ({
+  name,
+  value,
+}: {
   name: string,
-  /* value: string, */
-};
-const DisplayCellContent = (props: DisplayCellContentProps) => {
-  const [field] = useField(props);
+  value: number,
+}) => {
 
   // 粗利率のみ単位を%、その他は3桁ごとにカンマ+単位に円を付与する
-  const dispVal = (field.name === 'grossProfitMargin') ? (field.value + ' %')
-    : (field.value.toLocaleString() + '円');
+  const dispVal = (name.indexOf('grossProfitMargin')) !== -1 ? (value + ' %')
+    : (Math.round(value).toLocaleString() + '円');
 
   return (
     <Typography variant='body2' >
